Abstract the current tag renaming fixers
