+++ Symfony/CS/Fixer/Symfony/PhpdocTypeToVarFixer.php
@@ -11,40 +11,33 @@
 
 namespace Symfony\CS\Fixer\Symfony;
 
-use Symfony\CS\AbstractFixer;
-use Symfony\CS\DocBlock\DocBlock;
-use Symfony\CS\Tokenizer\Tokens;
+use Symfony\CS\AbstractPhpdocTagsFixer;
 
 /**
  * @author Graham Campbell <graham@mineuk.com>
  */
-class PhpdocTypeToVarFixer extends AbstractFixer
+class PhpdocTypeToVarFixer extends AbstractPhpdocTagsFixer
 {
     /**
-     * {@inheritdoc}
+     * The tag to search for.
+     *
+     * @var string
      */
-    public function fix(\SplFileInfo $file, $content)
-    {
-        $tokens = Tokens::fromCode($content);
+    protected static $search = 'type';
 
-        foreach ($tokens->findGivenKind(T_DOC_COMMENT) as $token) {
-            $doc = new DocBlock($token->getContent());
-            $annotations = $doc->getAnnotationsOfType('type');
+    /**
+     * The input tag.
+     *
+     * @var string
+     */
+    protected static $input = '@type';
 
-            if (empty($annotations)) {
-                continue;
-            }
-
-            foreach ($annotations as $annotation) {
-                $line = $doc->getLine($annotation->getStart());
-                $line->setContent(str_replace('@type', '@var', $line->getContent()));
-            }
-
-            $token->setContent($doc->getContent());
-        }
-
-        return $tokens->generateCode();
-    }
+    /**
+     * The output tag.
+     *
+     * @var string
+     */
+    protected static $output = '@var';
 
     /**
      * {@inheritdoc}
